+++ app/dashboard/properties/[propertyId]/events/add/_components/FormContent/ElectricityEventContent/ElectricityEventContent.tsx
@@ -1,8 +1,0 @@
-import { FormControl, Input, NullOption } from '@/components/UI/FormUtils';
-import { TargetSelector } from './TargetSelector';
-
-export const ElectricityEventContent = () => (
-  <>
-    <TargetSelector />
-  </>
-);